etl_nba_portafolio/src/requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
schedule
